requirements.txt
fastapi
uvicorn
mysql-connector-python
boto3
python-dotenv